move ui methods into app.ui
